mainwindow refactoring.
